refactor: extract shared Quantities and Ticks for both transports
- Move Quantities to src/quantities.rs, re-export from both inserters
- Make Ticks available to native transport (fallback to std::time::Instant
  when quanta/inserter feature not enabled)
- Remove NativeTicks duplication — native inserter now uses shared Ticks
- Net -38 lines of duplicated code
